# src/utils/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a API Rest for Actions in backend improve-in application.",
    "version": "1.0.0",
    "title": "Swagger improve-in",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "mauriciogallegoavi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api/person": {
      "get": {
        "tags": [
          "get"
        ],
        "summary": "get people",
        "description": "",
        "operationId": "getpeople",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/addPeople"
              }
            }
          },
          "400": {
            "description": "Invalid"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "password",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "addPeople": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "age":{
          "type": "integer",
          "format": "int64"
        },
        "genre": {
          "type": "string"
        }
      }
    },
...
  },
...
}